Update project for simple rest clients
JSON is not working so far. We need to find out why. For now, we can exemplify rest communication via values passed in the URL path instead.

In the future, the idea is to add an example with Query Params and another using JSON. By doing so, we will have the main three examples of how quarkus applications can communicate via REST